requirements.txt
Flask==3.0.2
mysql-connector-python==8.3.0
docx2pdf==0.1.8
pytesseract==0.3.10
Pillow==10.2.0
pywin32==306
Werkzeug==3.0.1
numpy==1.26.2
setuptools>=68.0.0
Flask-SocketIO 